+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 
 > The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.0.1
+- Added support for including multiple CSS and JS files at once
+- Added `s()` helper method
+
 ## 1.0.0
 - **Added Base Classes**
     - BaseCommand